Playing around with an idea for the bake operation
On jabbr had brief discussion about treating bake operation as
a pipeline of commands that are run through. Jekyll source uses
a Site which has read, generate, and write steps (plus a few others).

I thought this Site could be the context of the pipeline, which would
include posts, layouts, static files etc, and each pipeline stage
would work with that information.

And just to be corny and keep with the theme I named this common
context "Oven" to fit in with the baking metaphor.

For this spike I added a test to make sure all stages were baked in
order using the same oven. Should also test that we create the oven
with the right options, but wasn't entirely sure how that info was
going to be passed through yet.

I'm also intentionally avoiding IoC containers at this point. The
BakeOperation knows how to create each of its steps at the moment.
The Func<Oven> factory method is just for testing.